intel-x86.inc: BSPs should not change DISTRO_FEATURES
If a BSP changes a distro feature, then hash values can change.  If
you try to build a single shared distribution with multiple BSPs, problems
may occur as the different packages may be configured differently.

The standard DISTRO_FEATURES being defined are already part of the
existing wrlinux distribution(s).  'vfat' is not intended to be a distro
flag, but instead a MACHINE_FEATURES.  vfat is already listed in this
configurations MACHINE_FEATURES.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>