intel-x86: use systemd-boot as bootloader when GPLv3 was not allowed
Issue: LIN1019-16

grub-efi use GPLv3 which may not allowed by some user, if so,
hdd and wic image is not available, to aviod this issue, when
GPLv3 was not permited, use systemd-boot as the bootloader for
x86 board.

Because wic.bbclass add both grub-efi and systemd-boot as its depency,
we remove it when we use systemd-boot, since systemd-boot can fully provide
what wic.bbclass are required.

Signed-off-by: Liwei Song <liwei.song@windriver.com>