Add -msse4.2, and update gcc build to select optimized settings
Issue: LIN1018-3076

Explicitly enable SSE instruction set by adding "-msse4.2" in tune
files for intel-haswell-64, intel-skylake-64 and intel-skylake-avx512.

When building lib32-gcc-runtime, XCFLAGS from GCC will overwrite -march
setting in BSP tune file, which causes SSE instruction to be disabled.

We add a manual setting of arch/cpu to GCC builds to ensure that the
proper configuration is passed through and not overwritten with the default
486/686 config.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>