canned-wks: workaround IMA signature failure on wic image
Since commit [1] applied in oe-core, if `--exclude-path boot/'
used in wks partition, the wic tool will copy pseudo database (files.db)
to new location.

But as pseudo manual's BUGS [2]
...
The  pseudo database is not particularly robust in the face of whole
directory trees being moved, or changes in the underlying device and
inode numbers. It has a reasonable chance of recovering if only the
path or the device numbers have changed, but it is not particularly
designed to address this.
...
In other words, it means when wic copy pseudo database, pseudo may
not flush it from memory to filesystem. It is the root cause of IMA
signature failure

One workaround is not to use option --exclude-path in wks. In this
case, the side effect is in rootfs partition(/dev/sda3), files in
/boot are not clean up.

[1] https://git.yoctoproject.org/cgit/cgit.cgi/poky/commit/?id=90164f0f3d713de70842317869343dd4a3ef5140
[2] http://manpages.ubuntu.com/manpages/xenial/man1/pseudo.1.html

Issue: LINCD-1826

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>